Signing JAR moved from manifest to jarutil
Corresponding tests updated accordingly.
