+++ javatools/jarutil.py
@@ -21,9 +21,39 @@ Java archives
 :license: LGPL
 """
 
+import os
+from zipfile import ZipFile
+
+from javatools.manifest import Manifest, SignatureManifest
 
 __all__ = ( "cli_create_jar", "cli_check_jar", "cli_sign_jar",
             "cli_verify_jar", "cli", "main" )
+
+
+def private_key_type(private_key_file):
+    import subprocess
+    import re
+
+    algorithms = ("RSA", "DSA", "EC")
+    # Grepping for a string will work for PKCS8 keys, but not for PKCS1.
+    with open(private_key_file, "r") as f:
+        # We can't just take the first line. PKCS8 may have other headers.
+        for line in f:
+            for algorithm in algorithms:
+                if re.match("-----BEGIN %s PRIVATE KEY-----" % algorithm,
+                            line):
+                    return algorithm
+
+    # No luck.
+    # Anything less ugly and more efficient, but working with all key types??
+    # PyOpenssl has Pkey.type()...
+    with open(os.devnull, "wb") as DEVNULL:
+        for algorithm in algorithms:
+            if not subprocess.call(
+                    ["openssl", algorithm.lower(), "-in", private_key_file],
+                    stdout=DEVNULL, stderr=subprocess.STDOUT):
+                return algorithm
+    return None
 
 
 def cli_create_jar(options, paths):
@@ -31,17 +61,41 @@ def cli_create_jar(options, paths):
     return 0
 
 
-def cli_check_jar(options, jarfilename):
-    # TODO: Verify the MANIFEST.MF checksums of a JAR
+def cli_sign_jar(options, jar_file, cert_file, key_file, key_alias):
+    """
+    Signs the jar (almost) identically to jarsigner.
+    """
+
+    jar = ZipFile(jar_file, "a")
+    if not "META-INF/MANIFEST.MF" in jar.namelist():
+        print "META-INF/MANIFEST.MF not found in %s" % jar_file
+        return 1
+
+    sig_block_extension = private_key_type(key_file)
+    if sig_block_extension is None:
+        print "Cannot determine private key type (is it in PEM format?)"
+        return 1
+
+    mf = Manifest()
+    mf.parse(jar.read("META-INF/MANIFEST.MF"))
+
+    # create a signature manifest, and make it match the line separator
+    # style of the manifest it'll be digesting.
+    sf = SignatureManifest(linesep=mf.linesep)
+
+    sf_digest_algorithm = "SHA-256"
+    if options and options.digest:
+        sf_digest_algorithm = options.digest
+    sf.digest_manifest(mf, sf_digest_algorithm)
+    jar.writestr("META-INF/%s.SF" % key_alias, sf.get_data())
+
+    sig_digest_algorithm = sf_digest_algorithm  # No point to make it different
+    jar.writestr("META-INF/%s.%s" % (key_alias, sig_block_extension),
+        sf.get_signature(cert_file, key_file, sig_digest_algorithm))
+
     return 0
 
-
-def cli_sign_jar(options, jarfilename, certfile, keyfile, alias):
-    # TODO: sign a JAR and embed the signature entries
-    return 0
-
-
-def cli_verify_jar(options, jarfilename, keystore=None):
+def cli_verify_jar_signature(options, jarfilename, keystore=None):
     # TODO: verify the signature in a JAR matches that from a known
     # key
     return 0
@@ -56,7 +110,22 @@ def cli(parser, options, rest):
 def create_optparser():
     from optparse import OptionParser
 
-    parser = OptionParser("%prog [OPTIONS] JARFILE")
+    parser = OptionParser(usage="%prog COMMAND [OPTIONS] JARFILE [ARGUMENTS]")
+
+    parser.add_option("-c", "--create")
+    # mandatory arguments: path(s)
+    # options: ignore, manifest-digest
+
+    parser.add_option("-s", "--sign",
+                      help="sign the JAR file"
+                      " (must be followed with: "
+                      "certificate.pem, private_key.pem, key_alias)")
+    # mandatory arguments: certificate, key, alias
+    # TODO: options: signature-file-digest
+
+    parser.add_option("-v", "--verify")
+    # no mandatory arguments
+    # options: trusted-certificates-storage, trust-all
 
     return parser
